+++ src/plastimatch/bspline_mse.cxx
@@ -263,7 +263,7 @@ bspline_score_g_mse (
     float* m_img = (float*)moving->img;
     float* m_grad = (float*)moving_grad->img;
 
-    size_t idx_tile;
+    int idx_tile;
 
     Plm_timer timer;
